twitter-bokuyaba-wiki-helper: automatically translate citation
Add automatic click on the "Translate tweet"/"Translate post" link and
automatic replacement of the wikitext for cases when the translation is
loaded.  This extra handling is needed, because sometimes Google
Translate refuses to translate a tweet and nothing loads after clicking
the link.

While we're here, refactor the creation of HTML elements so that every
"create..." function returns the element and it is appended where needed
by the caller.  Rename functions that do the appending themselves to
"append...".